requirements.txt
certifi==2021.10.8
click==8.0.1
cycler==0.10.0
decorator==5.1.0
future==0.18.2
hydra-colorlog==1.1.0
hydra-core==1.1.0
hydra-optuna-sweeper==1.1.0
joblib==1.0.1
json5==0.9.6
jsonschema==4.2.1
kiwisolver==1.3.1
lightning-bolts==0.4.0
madgrad==1.1
matplotlib-inline==0.1.3
matplotlib==3.4.2
notebook==6.4.6
numpy==1.20.3
opencv-python==4.5.4.60
optuna==2.4.0
packaging==21.3
pandas==1.2.3
pathtools==0.1.2
pillow==8.4.0
pip==21.1.2
py==1.11.0
pyflakes==2.4.0
pygments==2.10.0
pylibjpeg-libjpeg==1.2.0
pylibjpeg-openjpeg==1.1.1
pylibjpeg-rle==1.1.0
pylibjpeg==1.3.0
pylint==2.10.2
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.1
python-dotenv==0.19.2
pytorch-lightning==1.5.5
rich==10.15.2
scikit-learn==0.24.1
scipy==1.6.3
seaborn==0.11.2]
six==1.16.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboard==2.5.0
threadpoolctl==2.1.0
torch==1.10.0
torchaudio==0.10.0
torchmetrics==0.6.1
torchvision==0.11.1
tqdm==4.61.1
typing-extensions==3.10.0.2
urllib3==1.26.5
virtualenv==20.10.0
wandb==0.12.7
wheel==0.36.2
yarl==1.7.2